fix: Runtime compilation errors
