+++ Assets/root/Runtime/Reflection/Serializers/RS_UnityEngineMaterial.Editor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1fdcf6b347a96b7428b817cbe47ff23b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
