+++ Assets/root/Runtime/Reflection/Serializers/RS_UnityEngineMaterial.Runtime.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2306b740284712d459ad1f2000fb006d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
