+++ Assets/root/Runtime/Reflection/Serializers/RS_UnityEngineObject.Editor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 47223c2894d564d45be4225f94de48cb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
