+++ Assets/root/Runtime/Reflection/Serializers/RS_UnityEngineObject.Runtime.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0d006ec32fcbb104cb373d02f05e054e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
