+++ Assets/root/Runtime/Reflection/Serializers/RS_UnityEngineSprite.Editor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b2c1483326912604cb8a5e0d0bb714c9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
