+++ Assets/root/Runtime/Reflection/Serializers/RS_UnityEngineSprite.Runtime.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6de533ec5d3f9a0488f8b2e7942f3af7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
